+++ familyhelper-webapi-impl/src/main/java/com/dwarfeng/familyhelper/webapi/impl/service/life/ActivityFileResponseServiceImpl.java
@@ -1,11 +1,15 @@
 package com.dwarfeng.familyhelper.webapi.impl.service.life;
 
-import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityFile;
-import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityFileUpdateInfo;
-import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityFileUploadInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.*;
 import com.dwarfeng.familyhelper.life.stack.bean.entity.ActivityFileInfo;
 import com.dwarfeng.familyhelper.life.stack.service.ActivityFileInfoMaintainService;
 import com.dwarfeng.familyhelper.life.stack.service.ActivityFileOperateService;
+import com.dwarfeng.familyhelper.plugin.commons.dto.VoucherIdWrapper;
+import com.dwarfeng.familyhelper.plugin.life.bean.dto.DubboRestActivityFileStream;
+import com.dwarfeng.familyhelper.plugin.life.bean.dto.DubboRestActivityFileStreamDownloadInfo;
+import com.dwarfeng.familyhelper.plugin.life.bean.dto.DubboRestActivityFileStreamUpdateInfo;
+import com.dwarfeng.familyhelper.plugin.life.bean.dto.DubboRestActivityFileStreamUploadInfo;
+import com.dwarfeng.familyhelper.plugin.life.service.DubboRestActivityFileOperateService;
 import com.dwarfeng.familyhelper.webapi.stack.service.life.ActivityFileResponseService;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
@@ -20,15 +24,19 @@ public class ActivityFileResponseServiceImpl implements ActivityFileResponseServ
 
     private final ActivityFileInfoMaintainService activityFileInfoMaintainService;
     private final ActivityFileOperateService activityFileOperateService;
+    private final DubboRestActivityFileOperateService dubboRestActivityFileOperateService;
 
     public ActivityFileResponseServiceImpl(
             @Qualifier("familyhelperLifeActivityFileInfoMaintainService")
             ActivityFileInfoMaintainService activityFileInfoMaintainService,
             @Qualifier("familyhelperLifeActivityFileOperateService")
-            ActivityFileOperateService activityFileOperateService
+            ActivityFileOperateService activityFileOperateService,
+            @Qualifier("familyhelperPluginLifeDubboRestActivityFileOperateService")
+            DubboRestActivityFileOperateService dubboRestActivityFileOperateService
     ) {
         this.activityFileInfoMaintainService = activityFileInfoMaintainService;
         this.activityFileOperateService = activityFileOperateService;
+        this.dubboRestActivityFileOperateService = dubboRestActivityFileOperateService;
     }
 
     @Override
@@ -104,6 +112,26 @@ public class ActivityFileResponseServiceImpl implements ActivityFileResponseServ
     }
 
     @Override
+    public LongIdKey requestActivityFileStreamVoucher(StringIdKey userKey, LongIdKey activityFileKey)
+            throws ServiceException {
+        VoucherIdWrapper voucherIdWrapper = dubboRestActivityFileOperateService.requestActivityFileStreamVoucher(
+                new DubboRestActivityFileStreamDownloadInfo(userKey.getStringId(), activityFileKey.getLongId())
+        );
+        return new LongIdKey(voucherIdWrapper.getVoucherId());
+    }
+
+    @Override
+    public ActivityFileStream downloadActivityFileStreamByVoucher(LongIdKey voucherKey) throws ServiceException {
+        VoucherIdWrapper voucherIdWrapper = new VoucherIdWrapper(voucherKey.getLongId());
+        DubboRestActivityFileStream dubboRestActivityFileStream =
+                dubboRestActivityFileOperateService.downloadActivityFileStreamByVoucher(voucherIdWrapper);
+        return new ActivityFileStream(
+                dubboRestActivityFileStream.getOriginName(), dubboRestActivityFileStream.getLength(),
+                dubboRestActivityFileStream.getContent()
+        );
+    }
+
+    @Override
     public void uploadActivityFile(
             StringIdKey userKey, ActivityFileUploadInfo activityFileUploadInfo
     ) throws ServiceException {
@@ -111,10 +139,30 @@ public class ActivityFileResponseServiceImpl implements ActivityFileResponseServ
     }
 
     @Override
+    public void uploadActivityFileStream(StringIdKey userKey, ActivityFileStreamUploadInfo activityFileStreamUploadInfo)
+            throws ServiceException {
+        dubboRestActivityFileOperateService.uploadActivityFileStream(new DubboRestActivityFileStreamUploadInfo(
+                userKey.getStringId(), activityFileStreamUploadInfo.getActivityKey().getLongId(),
+                activityFileStreamUploadInfo.getOriginName(), activityFileStreamUploadInfo.getLength(),
+                activityFileStreamUploadInfo.getContent()
+        ));
+    }
+
+    @Override
     public void updateActivityFile(
             StringIdKey userKey, ActivityFileUpdateInfo activityFileUpdateInfo
     ) throws ServiceException {
         activityFileOperateService.updateActivityFile(userKey, activityFileUpdateInfo);
+    }
+
+    @Override
+    public void updateActivityFileStream(StringIdKey userKey, ActivityFileStreamUpdateInfo activityFileStreamUpdateInfo)
+            throws ServiceException {
+        dubboRestActivityFileOperateService.updateActivityFileStream(new DubboRestActivityFileStreamUpdateInfo(
+                userKey.getStringId(), activityFileStreamUpdateInfo.getActivityFileKey().getLongId(),
+                activityFileStreamUpdateInfo.getOriginName(), activityFileStreamUpdateInfo.getLength(),
+                activityFileStreamUpdateInfo.getContent()
+        ));
     }
 
     @Override